src/test/java/org/example/restapi/note/NoteControllerTest.java
package org.example.restapi.note;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.example.restapi.note.request.CreateNoteRequest;
import org.example.restapi.note.response.CreateNoteResponse;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.security.test.context.support.WithMockUser;

import static org.junit.jupiter.api.Assertions.*;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.*;


@ExtendWith(SpringExtension.class)
@SpringBootTest
@AutoConfigureMockMvc
public class NoteControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private NoteServiceImpl noteService;
    @InjectMocks
    private NoteController noteController;
    @Autowired
    private ObjectMapper objectMapper;


    @BeforeEach
    public void setup() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    @WithMockUser(username = "testUser", password = "testPassword", roles = "USER")
    public void testCreateNote() throws Exception {
        CreateNoteRequest request = new CreateNoteRequest();
        request.setTitle("Test Title");
        request.setContent("Test Content");

        CreateNoteResponse expectedResponse = CreateNoteResponse.success(1L);

        when(noteService.create(anyString(), any(CreateNoteRequest.class))).thenReturn(expectedResponse);

        MvcResult result = mockMvc.perform(MockMvcRequestBuilders.post("/api/v1/notes/create")
                        .content(asJsonString(request))
                        .contentType(MediaType.APPLICATION_JSON)
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andReturn();

        String content = result.getResponse().getContentAsString();
        CreateNoteResponse actualResponse = objectMapper.readValue(content, CreateNoteResponse.class);

        assertEquals(expectedResponse, actualResponse);
        verify(noteService, times(1)).create(anyString(), any(CreateNoteRequest.class));
    }


    private String asJsonString(final Object obj) {
        try {
            return new ObjectMapper().writeValueAsString(obj);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
}
